refactor: Replace NextQuestion component by a conditional navigation
