--- app/views/screens/Quiz/NextQuestion/index.js
@@ -1,3 +1,0 @@
-import NextQuestion from './NextQuestion';
-
-export default NextQuestion;
