+++ app/views/screens/Quiz/Question/index.js
@@ -1,0 +1,3 @@
+import Question from './Question';
+
+export default Question;
