+++ app/views/screens/Quiz/Score/Score.js
@@ -1,13 +1,38 @@
 import React, { Component } from 'react';
 import { View, Text, StyleSheet } from 'react-native';
-import baseStyles, { colors } from '../../../styles';
 import { connect } from 'react-redux';
+import { quizActions } from 'state/quizzes';
+import { NavigationEvents } from 'react-navigation';
+import baseStyles, { colors } from '../../../styles';
+import { setQuizTitle } from 'utils/helpers';
 
 class Score extends Component {
+  static navigationOptions = setQuizTitle;
+
+  state = {
+    goBack: true
+  };
+
+  onWillBlur = () => {
+    const { decreaseStep, quizId } = this.props;
+
+    if (this.state.goBack) {
+      setTimeout(() => {
+        decreaseStep(quizId);
+      }, 50);
+    } else {
+      this.setState({
+        goBack: true
+      });
+    }
+  };
+
   render() {
     const { correct, total } = this.props;
+
     return (
       <View style={styles.container}>
+        <NavigationEvents onWillBlur={this.onWillBlur} />
         <View style={baseStyles.card}>
           <Text style={styles.header}>Your score</Text>
           <View style={styles.scoreContainer}>
@@ -53,9 +78,11 @@ const styles = StyleSheet.create({
   }
 });
 
-const mapStateToProps = ({ quizzes, decks }, { quizId }) => {
+const mapStateToProps = ({ quizzes, decks }, { navigation }) => {
+  const quizId = navigation.getParam('quizId');
   const quiz = quizzes[quizId];
   return {
+    quizId,
     correct: Object.values(quiz.results).reduce(
       (acc, result) => (result ? acc + 1 : acc),
       0
@@ -64,4 +91,11 @@ const mapStateToProps = ({ quizzes, decks }, { quizId }) => {
   };
 };
 
-export default connect(mapStateToProps)(Score);
+const mapDispatchToProps = {
+  decreaseStep: id => quizActions.returnQuiz(id)
+};
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(Score);
