+++ app/views/screens/Quiz/Score/index.js
@@ -1,0 +1,3 @@
+import Score from './Score';
+
+export default Score;
