+++ app/views/screens/Quiz/index.js
@@ -1,2 +1,3 @@
 export { default as QuizAnswer } from './Answer';
-export { default as QuizNextQuestion } from './NextQuestion';
+export { default as QuizQuestion } from './Question';
+export { default as QuizScore } from './Score';
