+++ app/views/screens/index.js
@@ -2,4 +2,4 @@ export { default as DeckDetail } from './DeckDetail';
 export { default as DeckList } from './DeckList';
 export { default as NewDeck } from './NewDeck';
 export { default as NewCard } from './NewCard';
-export { QuizAnswer, QuizNextQuestion } from './Quiz';
+export { QuizAnswer, QuizQuestion, QuizScore } from './Quiz';
